+++ blockstats.tcl
@@ -26,6 +26,8 @@
 proc block_info {nick host hand chan arg} {
 	global help_blocktime help_blocked channels debug debugoutput output onlyallowregisteredusers output_blockinfo output_blockinfo_percoin command_protect
 
+	if {$debug eq "1"} { putlog "running proc [dict get [info frame 0] proc]" }
+	
 	if {$onlyallowregisteredusers eq "1"} {
 		if {[check_registereduser $chan $nick] eq "false"} {
 			putquick "NOTICE $nick :you are not allowed to use this command"
@@ -142,6 +144,8 @@ proc block_info {nick host hand chan arg} {
 proc last_info {nick host hand chan arg } {
 	global help_blocktime help_blocked channels debug debugoutput output onlyallowregisteredusers output_lastblock output_lastblock_percoin command_protect
 
+	if {$debug eq "1"} { putlog "running proc [dict get [info frame 0] proc]" }
+	
 	if {$onlyallowregisteredusers eq "1"} {
 		if {[check_registereduser $chan $nick] eq "false"} {
 			putquick "NOTICE $nick :you are not allowed to use this command"