+++ marketdata.tcl
@@ -26,6 +26,8 @@
 proc price_info {nick host hand chan arg} {
 	global help_blocktime help_blocked channels debug debugoutput usehttps output onlyallowregisteredusers output_marketdata command_protect marketapi_coinse marketapi_vircurex marketapi_mintpal
 
+	if {$debug eq "1"} { putlog "running proc [dict get [info frame 0] proc]" }
+	
 	if {$onlyallowregisteredusers eq "1"} {
 		if {[check_registereduser $chan $nick] eq "false"} {
 			putquick "NOTICE $nick :you are not allowed to use this command"
@@ -65,19 +67,15 @@ proc price_info {nick host hand chan arg} {
 	set trade_label "0"
 	set trade_volume "0"
 
-	if {$query_exchange eq "COINS-E"} {
-		set market_name "Coins-E"
-		set newurl $marketapi_coinse
-		putlog "URL: $newurl"
-	} elseif {$query_exchange eq "VIRCUREX"} {
+	if {$query_exchange eq "VIRCUREX"} {
 		set market_name "Vircurex"
 		set newurl $marketapi_vircurex
 		append newurl "?base=$query_basecoin&alt=$query_altcoin"
 		putlog "URL: $newurl"
-	} elseif {$query_exchange eq "MINTPAL"} {
-		set market_name "MintPal"
-		set newurl $marketapi_mintpal
-		append newurl "$query_altcoin/$query_basecoin"
+	} elseif {$query_exchange eq "CRYPTOPIA"} {
+		set market_name "Cryptopia"
+		set newurl $marketapi_cryptopia
+		append newurl "?base=$query_basecoin&alt=$query_altcoin"
 		putlog "URL: $newurl"
 	} else {
 		if {$output eq "CHAN"} {
@@ -118,86 +116,12 @@ proc price_info {nick host hand chan arg} {
 }
 
 #
-# output for coins-e api
-#
-proc market_coinse {nick chan marketdataresult query_altcoin query_basecoin} {
-	global debug debugoutput output output_marketdata_coinse
-	
-	set querycoinpair "$query_altcoin"
-	append querycoinpair "_"
-	append querycoinpair "$query_basecoin"
-	
-	if {$debug eq "1"} { putlog "QUERY: $querycoinpair" }
-
-	set trade_volume "null"
-	set trade_high "null"
-	set trade_low "null"
-	set trade_avg "null"
-	
-	foreach {key value} $marketdataresult {
-		#putlog "DATA: $key"
-		if {$key eq "markets"} {
-			foreach {sub_key sub_value} $value {
-				#putlog "DATA: $sub_key"
-				if {$sub_key eq "$querycoinpair"} {
-					#putlog "DATA: $sub_value"
-					foreach {elem elem_val} $sub_value {
-						#putlog "DATA: $elem"
-						if {$elem eq "c2"} {
-							set basecoin $elem_val
-						}
-						if {$elem eq "c1"} {
-							set altcoin "$elem_val"
-						}						
-
-						foreach {elem2 elem_val2} $elem_val {
-							#putlog "DATA: $elem2"
-
-							if {$elem2 eq "24h"} {
-								foreach {elem3 elem_val3} $elem_val2 {
-									#putlog "DATA: $elem3 - $elem_val3"
-									if {$elem3 eq "volume"} { set trade_volume "$elem_val3" }
-									if {$elem3 eq "h"} { set trade_high "$elem_val3" }
-									if {$elem3 eq "l"} { set trade_low "$elem_val3" }
-									if {$elem3 eq "avg_rate"} { set trade_avg "$elem_val3" }
-								}
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-
-	if {$trade_volume eq "null" } { 
-		putquick "PRIVMSG $chan :Unknown currency pair"
-		return
-	}
-
-	set lineoutput $output_marketdata_coinse
-	set lineoutput [replacevar $lineoutput "%marketdata_market%" "Coins-E"]
-	set lineoutput [replacevar $lineoutput "%marketdata_trade_basecoin%" $query_basecoin]
-	set lineoutput [replacevar $lineoutput "%marketdata_trade_altcoin%" $query_altcoin]
-	set lineoutput [replacevar $lineoutput "%marketdata_tradehigh%" $trade_high]
-	set lineoutput [replacevar $lineoutput "%marketdata_tradelow%" $trade_low]
-	set lineoutput [replacevar $lineoutput "%marketdata_tradeavg%" $trade_avg]
-	set lineoutput [replacevar $lineoutput "%marketdata_tradevolume%" $trade_volume]
-	set lineoutput [replacevar $lineoutput "%marketdata_basecoin%" $basecoin]
-
-	if {$output eq "CHAN"} {
-		putquick "PRIVMSG $chan :$lineoutput"
-	} elseif {$output eq "NOTICE"} {
-		putquick "NOTICE $nick :$lineoutput"
-	} else {
-		putquick "PRIVMSG $chan :please set output in config file"
-	}
-}
-
-#
 # output for vircurex api
 #
 proc market_vircurex {nick chan marketdataresult query_altcoin query_basecoin} {
 	global debug debugoutput output output_marketdata_vircurex
+	
+	if {$debug eq "1"} { putlog "running proc [dict get [info frame 0] proc]" }
 	
 	if {$debug eq "1"} { putlog "QUERY: $query_altcoin\/$query_basecoin" }
 
@@ -236,10 +160,12 @@ proc market_vircurex {nick chan marketdataresult query_altcoin query_basecoin} {
 }
 
 #
-# output for mintpal api
+# output for cryptopia api
 #
-proc market_mintpal {nick chan marketdataresult query_altcoin query_basecoin} {
+proc market_crptopia {nick chan marketdataresult query_altcoin query_basecoin} {
 	global debug debugoutput output output_marketdata_mintpal
+	
+	if {$debug eq "1"} { putlog "running proc [dict get [info frame 0] proc]" }
 	
 	if {$debug eq "1"} { putlog "QUERY: $query_altcoin\/$query_basecoin" }