+++ workers.tcl
@@ -25,6 +25,8 @@
 #
 proc worker_info {nick host hand chan arg} {
 	global help_blocktime help_blocked channels debug debugoutput output onlyallowregisteredusers ownersworkeronly output_workerinfo output_worker_online output_worker_offline command_protect
+
+	if {$debug eq "1"} { putlog "running proc [dict get [info frame 0] proc]" }
 	
 	# only allow bot owners to get workers for 
 	# specified users